Update to use Bun instead of Node and better-sqlite-3

+++ sites/www/src/lib/server/db/interactions.ts
@@ -11,15 +11,19 @@ export function get_user_likes_and_saves(
 	const user_likes = new Set<number>()
 	const user_saves = new Set<number>()
 
-	const likeStmt = db.prepare('SELECT 1 FROM likes WHERE user_id = ? AND target_id = ?')
-	const saveStmt = db.prepare('SELECT 1 FROM saves WHERE user_id = ? AND target_id = ?')
+	const likeStmt = db.prepare(
+		'SELECT 1 FROM likes WHERE user_id = $user_id AND target_id = $target_id'
+	)
+	const saveStmt = db.prepare(
+		'SELECT 1 FROM saves WHERE user_id = $user_id AND target_id = $target_id'
+	)
 
 	db.transaction(() => {
 		for (const content_id of content_ids) {
-			if (likeStmt.get(user_id, content_id)) {
+			if (likeStmt.get({ $user_id: user_id, $target_id: content_id })) {
 				user_likes.add(content_id)
 			}
-			if (saveStmt.get(user_id, content_id)) {
+			if (saveStmt.get({ $user_id: user_id, $target_id: content_id })) {
 				user_saves.add(content_id)
 			}
 		}
@@ -39,12 +43,12 @@ export function get_user_likes_and_saves_count(user_id: number | undefined): {
 	let user_likes = 0
 	let user_saves = 0
 
-	const likeStmt = db.prepare('SELECT COUNT(1) AS count FROM likes WHERE user_id = ?')
-	const saveStmt = db.prepare('SELECT COUNT(1) AS count FROM saves WHERE user_id = ?')
+	const likeStmt = db.prepare('SELECT COUNT(1) AS count FROM likes WHERE user_id = $id')
+	const saveStmt = db.prepare('SELECT COUNT(1) AS count FROM saves WHERE user_id = $id')
 
 	db.transaction(() => {
-		user_likes = (likeStmt.get(user_id) as { count: number }).count
-		user_saves = (saveStmt.get(user_id) as { count: number }).count
+		user_likes = (likeStmt.get({ id: user_id }) as { count: number }).count
+		user_saves = (saveStmt.get({ id: user_id }) as { count: number }).count
 	})()
 
 	return { user_likes, user_saves }
@@ -52,14 +56,15 @@ export function get_user_likes_and_saves_count(user_id: number | undefined): {
 
 type InteractionType = 'like' | 'save'
 
-export function add_interaction(type: InteractionType, userId: string, contentId: string): void {
-	const table = `${type}s`
-	const query = `INSERT OR IGNORE INTO ${table} (user_id, target_id, created_at) VALUES (?, ?, CURRENT_TIMESTAMP)`
-	db.prepare(query).run(userId, contentId)
+export function add_interaction(type: InteractionType, user_id: string, contentId: string): void {
+	const query = db.prepare(
+		`INSERT OR IGNORE INTO ${type}s (user_id, target_id, created_at) VALUES ($user_id, $target_id, CURRENT_TIMESTAMP)`
+	)
+	query.run({ user_id, target_id: contentId })
 }
 
 export function remove_interaction(type: InteractionType, userId: string, contentId: string): void {
 	const table = `${type}s`
-	const query = `DELETE FROM ${table} WHERE user_id = ? AND target_id = ?`
-	db.prepare(query).run(userId, contentId)
+	const query = `DELETE FROM ${table} WHERE user_id = $user_id AND target_id = $target_id`
+	db.prepare(query).run({ user_id: userId, target_id: contentId })
 }
